Move context provide up one layer to the DevIdePage
Related to #360
