+++ app/web/src/pages/DevIdePage/DevIdePage.js
@@ -1,8 +1,12 @@
+import { createContext } from 'react'
 import Seo from 'src/components/Seo/Seo'
 import IdeToolbar from 'src/components/IdeToolbarNew'
 import { Toaster } from '@redwoodjs/web/toast'
+import { useIdeState } from 'src/helpers/hooks/useIdeState'
 
+export const IdeContext = createContext()
 const DevIdePage = ({ cadPackage }) => {
+  const [state, thunkDispatch] = useIdeState()
   return (
     <div className="h-screen flex flex-col">
       <Seo
@@ -11,7 +15,9 @@ const DevIdePage = ({ cadPackage }) => {
         lang="en-US"
       />
       <Toaster timeout={9000} />
-      <IdeToolbar cadPackage={cadPackage} />
+      <IdeContext.Provider value={{ state, thunkDispatch }}>
+        <IdeToolbar cadPackage={cadPackage} />
+      </IdeContext.Provider>
     </div>
   )
 }
